backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
websockets==12.0
aiofiles==24.1.0
psutil==5.9.8
docker==6.1.3
requests==2.31.0
urllib3==1.26.18
gputil==1.4.0
pydantic==2.6.1
python-multipart==0.0.9
python-dateutil==2.8.2
httpx==0.27.0
pyyaml==6.0.1
bcrypt==4.0.1
PyJWT==2.8.0
redis==5.0.1
hiredis==2.3.2
cryptography==42.0.5
itsdangerous==2.1.2
stripe==10.0.0
alembic==1.13.1
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0
apscheduler==3.10.4
boto3==1.34.0
Jinja2==3.1.2
aiosmtplib==3.0.2
msal==1.26.0
email-validator==2.1.0
numpy==1.24.0
scipy==1.11.0
Pillow==10.2.0

# OpenTelemetry for distributed tracing
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-httpx==0.42b0
opentelemetry-instrumentation-asyncpg==0.42b0
opentelemetry-instrumentation-redis==0.42b0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
opentelemetry-exporter-jaeger==1.21.0

# Testing dependencies
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0
pytest-mock==3.12.0
slowapi==0.1.9
